requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dateutils==0.6.12
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.11.0
django-environ==0.9.0
django-heroku==0.3.1
django-rest-framework-braces==0.3.4
djangorestframework==3.13.1
djangorestframework-gis==0.18
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
ez-setup==0.9
feedparser==6.0.10
filelock==3.7.1
idna==3.4
MarkupSafe==2.1.1
psycopg2==2.9.3
PyJWT==2.5.0
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
stripogram==1.5
urllib3==1.26.12
Werkzeug==2.2.2
whitenoise==6.2.0
dotenv
